Create latents in float32
The denoising loop always computes the next step in float32, so this
would fail when using `bfloat16`.

+++ src/diffusers/pipelines/stable_diffusion/pipeline_flax_stable_diffusion.py
@@ -157,7 +157,7 @@ class FlaxStableDiffusionPipeline(FlaxDiffusionPipeline):
             self.unet.sample_size,
         )
         if latents is None:
-            latents = jax.random.normal(prng_seed, shape=latents_shape, dtype=self.dtype)
+            latents = jax.random.normal(prng_seed, shape=latents_shape, dtype=jnp.float32)
         else:
             if latents.shape != latents_shape:
                 raise ValueError(f"Unexpected latents shape, got {latents.shape}, expected {latents_shape}")
